examples/phone-chatbot/daily-twilio-sip-dial-in/requirements.txt
pipecat-ai[daily,cartesia,openai,silero]
fastapi==0.115.6
uvicorn
python-dotenv
twilio
aiohttp
googlemaps
boto3
